feat: Allow '.md' extension
This makes it easier for text editors and github to format the files for
human reading / editing.

In doing this, we need to gracefully handle unrelated `.md` files.  This
means we allow, rather than error, if there are no test cases.
We don't want syntax mistakes to cause tests to be silently ignored, so
we elevate these cases as `ignored` in the UI.