fix!: Don't fail on Linux with `[EXE]`
On Windows, we'd turn `[EXE]` into `.exe`, match, and all is well.  On
Linux, the pattern will have `[EXE]` but it becomes an unknown variable,
so we treat it as a literal which won't exist in the line, and the test
fails.  So let's wipe these blank variables from the pattern.

BREAKING CHANGE: `Substitutions` trait has a new `clear` method.