feat(md)!: Allow verifying without trailing newline
Because of how we parsed the file, the stdout for an `md` code block
always had a newline at the end.  This prevents verifying applications
that do not output a trailing newline (usually applications should).

By stripping and re-adding the newline, we also provide more visual
separarion between the commands, making the output eaiser to read.

BREAKING CHANGE: All `md` command outputs need an extra newline