fix: Auto-detect binary files
`binary` only impacts stdin/stdout/stderr.  File systems tend to be too
diverse to be controlled by a single flag.  Before we erred on the side
of always assuming text.  Now we will assume a file is text only if it
doesn't look like its binary and it is valid UTF-8

Fixes #61