feat(assert): Allow disabling path normalization
This handles the snapbox side of #91.  We'll need to update the plumbing
in trycmd to get this to all the right places.  For now this also lets
us experiment with it where we can more easily change the API.

We still normalize paths when inserting substitions.  The two APIs
aren't connected at the moment to allow fixing this.  This will need
more consideration.